# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backcast EVS",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/cost-registrations": {
      "get": {
        "tags": [
          "Cost Registrations"
        ],
        "summary": "Read Cost Registrations",
        "description": "Retrieve cost registrations with server-side search, filtering, and sorting.\n\nCost registrations track actual expenditures against cost elements.\nThey are versionable but NOT branchable (costs are global facts).",
        "operationId": "get_cost_registrations",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number (1-indexed)",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number (1-indexed)"
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Per Page"
            },
            "description": "Items per page"
          },
          {
            "name": "cost_element_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by Cost Element ID",
              "title": "Cost Element Id"
            },
            "description": "Filter by Cost Element ID"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search term (description, invoice, vendor)",
              "title": "Search"
            },
            "description": "Search term (description, invoice, vendor)"
          },
          {
            "name": "filters",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filters in format 'column:value;column:value1,value2'",
              "title": "Filters"
            },
            "description": "Filters in format 'column:value;column:value1,value2'"
          },
          {
            "name": "sort_field",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Field to sort by",
              "title": "Sort Field"
            },
            "description": "Field to sort by"
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "description": "Sort order (asc or desc)",
              "default": "asc",
              "title": "Sort Order"
            },
            "description": "Sort order (asc or desc)"
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Time travel: get Cost Registrations as of this timestamp (ISO 8601)",
              "title": "As Of"
            },
            "description": "Time travel: get Cost Registrations as of this timestamp (ISO 8601)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Cost Registrations"
        ],
        "summary": "Create Cost Registration",
        "description": "Create a new cost registration.\n\nValidates that the cost does not exceed the cost element's budget.\nRaises BudgetExceededError if budget would be exceeded.",
        "operationId": "create_cost_registration",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CostRegistrationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CostRegistrationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cost-registrations/{cost_registration_id}": {
      "get": {
        "tags": [
          "Cost Registrations"
        ],
        "summary": "Read Cost Registration",
        "description": "Get a specific cost registration by id.\n\nSupports time-travel queries via the as_of parameter to view\nthe cost registration's state at any historical point in time.",
        "operationId": "get_cost_registration",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cost_registration_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cost Registration Id"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Time travel: get cost registration state as of this timestamp (ISO 8601)",
              "title": "As Of"
            },
            "description": "Time travel: get cost registration state as of this timestamp (ISO 8601)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CostRegistrationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Cost Registrations"
        ],
        "summary": "Update Cost Registration",
        "description": "Update a cost registration.\n\nCreates a new version of the cost registration with the updated values.\nPrevious versions are preserved in the history.",
        "operationId": "update_cost_registration",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cost_registration_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cost Registration Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CostRegistrationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CostRegistrationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Cost Registrations"
        ],
        "summary": "Delete Cost Registration",
        "description": "Soft delete a cost registration.\n\nMarks the cost registration as deleted but preserves it in the history.",
        "operationId": "delete_cost_registration",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cost_registration_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cost Registration Id"
            }
          },
          {
            "name": "control_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional control date for deletion",
              "title": "Control Date"
            },
            "description": "Optional control date for deletion"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cost-registrations/{cost_registration_id}/history": {
      "get": {
        "tags": [
          "Cost Registrations"
        ],
        "summary": "Get Cost Registration History",
        "description": "Get full version history for a cost registration.\n\nReturns all versions of the cost registration, ordered by transaction time.\nIncludes both current and historical versions.",
        "operationId": "get_cost_registration_history",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cost_registration_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cost Registration Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CostRegistrationRead"
                  },
                  "title": "Response Get Cost Registration History"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cost-registrations/budget-status/{cost_element_id}": {
      "get": {
        "tags": [
          "Cost Registrations"
        ],
        "summary": "Get Budget Status",
        "description": "Get budget status for a cost element.\n\nReturns the budget amount, used amount, remaining amount, and percentage used.\nUseful for displaying budget progress bars and warnings.",
        "operationId": "get_budget_status",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cost_element_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cost Element Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Budget Status"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CostRegistrationCreate": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,2}0*$"
              }
            ],
            "title": "Amount",
            "description": "Cost amount (must be positive)"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,2}0*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity",
            "description": "Quantity of units consumed (optional)"
          },
          "unit_of_measure": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Of Measure",
            "description": "Unit of measure (e.g., 'hours', 'kg', 'm', 'each')"
          },
          "registration_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registration Date",
            "description": "When the cost was incurred (defaults to control date if not provided)"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Optional description of the cost"
          },
          "invoice_number": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Invoice Number",
            "description": "Optional invoice reference"
          },
          "vendor_reference": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor Reference",
            "description": "Optional vendor/supplier reference"
          },
          "cost_registration_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Registration Id",
            "description": "Root Cost Registration ID (internal use only for seeding)"
          },
          "cost_element_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cost Element Id",
            "description": "ID of the cost element to charge"
          }
        },
        "type": "object",
        "required": [
          "amount",
          "cost_element_id"
        ],
        "title": "CostRegistrationCreate",
        "description": "Properties required for creating a Cost Registration."
      },
      "CostRegistrationRead": {
        "properties": {
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,2}0*$",
            "title": "Amount",
            "description": "Cost amount (must be positive)"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,2}0*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity",
            "description": "Quantity of units consumed (optional)"
          },
          "unit_of_measure": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Of Measure",
            "description": "Unit of measure (e.g., 'hours', 'kg', 'm', 'each')"
          },
          "registration_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registration Date",
            "description": "When the cost was incurred (defaults to control date if not provided)"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Optional description of the cost"
          },
          "invoice_number": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Invoice Number",
            "description": "Optional invoice reference"
          },
          "vendor_reference": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor Reference",
            "description": "Optional vendor/supplier reference"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "cost_registration_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cost Registration Id"
          },
          "cost_element_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cost Element Id"
          },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By"
          }
        },
        "type": "object",
        "required": [
          "amount",
          "id",
          "cost_registration_id",
          "cost_element_id",
          "created_by"
        ],
        "title": "CostRegistrationRead",
        "description": "Properties returned to client."
      },
      "CostRegistrationUpdate": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,2}0*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,2}0*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "unit_of_measure": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Of Measure"
          },
          "registration_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registration Date"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "invoice_number": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Invoice Number"
          },
          "vendor_reference": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor Reference"
          }
        },
        "type": "object",
        "title": "CostRegistrationUpdate",
        "description": "Properties that can be updated on a Cost Registration."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}